refactor: 로그인/로그아웃 기능 드롭다운으로 이동
